+++ nuclia/lib/kb.py
@@ -34,6 +34,7 @@ TUS_UPLOAD_RESOURCE_URL = "/resource/{rid}/file/{field}/tusupload"
 TUS_UPLOAD_URL = "/tusupload"
 ACTIVITY_LOG_URL = "/activity/download?type={type}&month={month}"
 FEEDBACK_LOG_URL = "/feedback/{month}"
+NOTIFICATIONS = "/notifications"
 
 
 class Environment(str, Enum):
@@ -171,6 +172,14 @@ class NucliaDBClient(BaseNucliaDBClient):
     def __repr__(self):
         return f"{self.environment} - {self.url}"
 
+    def notifications(self):
+        if self.url is None or self.stream_session is None:
+            raise Exception("KB not configured")
+        url = f"{self.url}{NOTIFICATIONS}"
+        response = self.stream_session.get(url, stream=True, timeout=3660)
+        handle_http_errors(response)
+        return response
+
     def ask(self, request: AskRequest, timeout: int = 1000):
         if self.url is None or self.stream_session is None:
             raise Exception("KB not configured")
@@ -230,9 +239,9 @@ class NucliaDBClient(BaseNucliaDBClient):
             "content-type": content_type,
         }
         if md5 is not None:
-            headers["upload-metadata"] += (
-                f",md5 {base64.b64encode(md5.encode()).decode()}"
-            )
+            headers[
+                "upload-metadata"
+            ] += f",md5 {base64.b64encode(md5.encode()).decode()}"
 
         response: httpx.Response = self.writer_session.post(url, headers=headers)
         handle_http_errors(response)
@@ -323,6 +332,15 @@ class AsyncNucliaDBClient(BaseNucliaDBClient):
                 base_url=url,  # type: ignore
             )
 
+    async def notifications(self):
+        if self.url is None or self.reader_session is None:
+            raise Exception("KB not configured")
+        url = f"{self.url}{NOTIFICATIONS}"
+        req = self.reader_session.build_request("GET", url, timeout=3660)
+        response = await self.reader_session.send(req, stream=True)
+        handle_http_errors(response)
+        return response
+
     async def ask(self, request: AskRequest, timeout: int = 1000):
         if self.url is None or self.reader_session is None:
             raise Exception("KB not configured")
@@ -404,9 +422,9 @@ class AsyncNucliaDBClient(BaseNucliaDBClient):
             "content-type": content_type,
         }
         if md5 is not None:
-            headers["upload-metadata"] += (
-                f",md5 {base64.b64encode(md5.encode()).decode()}"
-            )
+            headers[
+                "upload-metadata"
+            ] += f",md5 {base64.b64encode(md5.encode()).decode()}"
 
         response = await self.writer_session.post(url, headers=headers)
         handle_http_errors(response)