+++ nuclia/sdk/auth.py
@@ -350,7 +350,7 @@ class NucliaAuth(BaseNucliaAuth):
             return resp.json()
         elif resp.status_code >= 300 and resp.status_code < 400:
             return None
-        elif resp.status_code == 403:
+        elif resp.status_code == 403 or resp.status_code == 401:
             raise UserTokenExpired()
         else:
             raise Exception({"status": resp.status_code, "message": resp.text})