+++ nuclia/tests/test_nua/test_agent.py
@@ -6,6 +6,6 @@ def test_agent(testing_config):
     agent_prompt = np.generate_prompt(
         text="You will answer questions about this topic in the way of a Spanish historian",
         agent_definition="Toledo",
-        model="anthropic",
+        model="chatgpt4o",
     )
     assert "Toledo" in agent_prompt